requirements.txt
gradio
opencv-python
numpy
pillow
